[RATOM-223] deselect messages after query, also move the export button again
