+++ src/components/Containers/Messages/MessagesContent/MessagesList/MessagesActions/MessagesActions.js
@@ -1,36 +1,16 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import {
-  borderSeparator,
-  colorGreyLight,
-  colorPrimary
-} from '../../../../../../styles/styleVariables';
+import { borderSeparator, colorGreyLight } from '../../../../../../styles/styleVariables';
 
 // Children
 import SelectionActions from './SelectionActions';
-import AccountExportModal from '../../../AccountExportModal';
-import Button from '../../../../../Components/Buttons/Button';
 
 const MessagesActions = () => {
-  const [showExportModal, setShowExportModal] = useState(false);
   return (
     <>
       <MessagesActionsStyled>
         <SelectionActions />
-        <Actions>
-          <Button
-            neutral
-            onClick={() => setShowExportModal(true)}
-            data-cy="export-as-records-request-button"
-          >
-            Export
-          </Button>
-        </Actions>
       </MessagesActionsStyled>
-      <AccountExportModal
-        isVisible={showExportModal}
-        closeModal={() => setShowExportModal(false)}
-      />
     </>
   );
 };
@@ -46,16 +26,6 @@ const MessagesActionsStyled = styled.div`
   display: flex;
   flex-direction: row;
   align-items: center;
-`;
-
-const Actions = styled.div`
-  max-width: 25%;
-
-  p {
-    font-size: 1.5rem;
-    cursor: pointer;
-    color: ${colorPrimary};
-  }
 `;
 
 export default MessagesActions;
