+++ src/components/Containers/Messages/MessagesContent/ResultsSummary.js
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React, { useState, useContext } from 'react';
 import styled from 'styled-components';
 import { borderSeparator, standardPadding, colorPrimary } from '../../../../styles/styleVariables';
 
@@ -6,19 +6,33 @@ import { borderSeparator, standardPadding, colorPrimary } from '../../../../styl
 import { AccountContext } from '../MessagesMain';
 import formatNumber from '../../../../util/formatNumber';
 
+// Children
+import AccountExportModal from '../AccountExportModal';
+
 const ResultsSummary = () => {
   const { messages, messagesTotalCount } = useContext(AccountContext);
-
+  const [showExportModal, setShowExportModal] = useState(false);
   return (
-    <ResultsSummaryStyled>
-      <Summary>
-        <h3>
-          Displaying <span>{formatNumber(messages.length)}</span> of
-          {messagesTotalCount === 10000 ? ' more than ' : ' '}
-          {formatNumber(messagesTotalCount)} results
-        </h3>
-      </Summary>
-    </ResultsSummaryStyled>
+    <>
+      <ResultsSummaryStyled>
+        <Summary>
+          <h3>
+            Displaying <span>{formatNumber(messages.length)}</span> of
+            {messagesTotalCount === 10000 ? ' more than ' : ' '}
+            {formatNumber(messagesTotalCount)} results
+          </h3>
+        </Summary>
+        <Actions>
+          <p onClick={() => setShowExportModal(true)} data-cy="export-as-records-request-button">
+            Export as Records Request
+          </p>
+        </Actions>
+      </ResultsSummaryStyled>
+      <AccountExportModal
+        isVisible={showExportModal}
+        closeModal={() => setShowExportModal(false)}
+      />
+    </>
   );
 };
 
@@ -40,6 +54,16 @@ const Summary = styled.div`
     span {
       color: ${colorPrimary};
     }
+  }
+`;
+
+const Actions = styled.div`
+  max-width: 25%;
+
+  p {
+    font-size: 1.5rem;
+    cursor: pointer;
+    color: ${colorPrimary};
   }
 `;
